extract modal to add new project to its own component
